Sanitize Maildir time stamps
The maildir storage plugin uses the time(2)/gettimeofday(2) based time stamp
for comparing messages' time stamps. It used to do this in an unsafe manner,
i.e. by using the file name's first component which contains the file name as
an input to int(). This commit adds sanitization that ensures there are only
digits in the time stamp (some programs writing to maildirs may add an
underscore to distinguish identical timestamps, producing something along the
lines of '123_0' and '123_1').