Message-ID based article retrieval
This commit implements message ID based article retrieval for the maildir
storage backend. Beyond the changes to the maildir backend it also required
some tweaks to papercut_nntp.py. Namely, a check for a storage class' backend
capabilities attribute was introduced in get_number_from_msg_id(). This kind of
attribute is transparent: storage backends can have one but they do not need
it. It is a dict indicating some capabilities not neccessarily related to
supported commands (such as the ability to handle message IDs as they are found
in an article's headers).